Move get_all_subclasses into commons
