+++ pyproject.toml
@@ -57,7 +57,6 @@ py-modules = [
     "matrix_functions",
     "matrix_functions_types",
     "optimizer_modules",
-    "utils",
 ]
 
 [tool.setuptools.packages.find]
